add a "config/helpers.py" and rearrange code
