+++ src/fides/api/ops/util/saas_util.py
@@ -14,7 +14,7 @@ from fides.api.ops.common_exceptions import FidesopsException
 from fides.api.ops.graph.config import Collection, CollectionAddress, Dataset, Field
 from fides.api.ops.schemas.saas.saas_config import SaaSRequest
 from fides.api.ops.schemas.saas.shared_schemas import SaaSRequestParams
-from fides.ctl.core.config.utils import load_file
+from fides.ctl.core.config.helpers import load_file
 from fides.lib.cryptography.cryptographic_util import bytes_to_b64_str
 
 logger = logging.getLogger(__name__)
