+++ src/fides/ctl/core/config/utils.py
@@ -18,65 +18,43 @@ def get_test_mode() -> bool:
     return test_mode
 
 
-def load_file(file_names: Union[List[Path], List[str]]) -> str:
-    """Load a file from the first matching location.
-
-    In order, will check:
-    - A path set at ENV variable FIDES__CONFIG_PATH
-    - The current directory
-    - The parent directory
-    - Two directories up for the current directory
-    - The parent_directory/.fides
-    - users home (~) directory
-    raises FileNotFound if none is found
-    """
-
-    possible_directories = [
-        os.getenv("FIDES__CONFIG_PATH"),
-        os.curdir,
-        os.pardir,
-        os.path.abspath(os.path.join(os.curdir, "..", "..")),
-        os.path.join(os.pardir, ".fides"),
-        os.path.expanduser("~"),
-    ]
-
-    directories = [d for d in possible_directories if d]
-
-    for dir_str in directories:
-        for file_name in file_names:
-            possible_location = os.path.join(dir_str, file_name)
-            if possible_location and os.path.isfile(possible_location):
-                logger.info("Loading file %s from %s", file_name, dir_str)
-                return possible_location
-
-    raise FileNotFoundError
-
-
-def get_config_from_file(
-    config_path_override: str,
-    section: str,
-    option: str,
-) -> Union[str, bool, int, None]:
-    """
-    Return the value currently written to the config file for the specified `section.option`.
-    """
-
-    try:
-        config_path = config_path_override or load_file(
-            file_names=[DEFAULT_CONFIG_PATH]
-        )
-    except FileNotFoundError as error:
-        raise error
-
-    current_config = load(config_path)
-
-    config_section = current_config.get(section)
-    if config_section is not None:
-        config_option = config_section.get(option)
-        if config_option is not None:
-            return config_option
-
-    return None
+CONFIG_KEY_ALLOWLIST = {
+    "cli": ["server_host", "server_port"],
+    "user": ["analytics_opt_out"],
+    "logging": ["level"],
+    "database": [
+        "server",
+        "user",
+        "port",
+        "db",
+        "test_db",
+    ],
+    "notifications": [
+        "send_request_completion_notification",
+        "send_request_receipt_notification",
+        "send_request_review_notification",
+        "notification_service_type",
+    ],
+    "redis": [
+        "host",
+        "port",
+        "charset",
+        "decode_responses",
+        "default_ttl_seconds",
+        "db_index",
+    ],
+    "security": [
+        "cors_origins",
+        "encoding",
+        "oauth_access_token_expire_minutes",
+    ],
+    "execution": [
+        "task_retry_count",
+        "task_retry_delay",
+        "task_retry_backoff",
+        "require_manual_request_approval",
+    ],
+}
 
 
 def check_required_webserver_config_values() -> None:
@@ -123,39 +101,3 @@ def check_required_webserver_config_values() -> None:
         )
 
         raise SystemExit(1)
-
-
-def update_config_file(  # type: ignore
-    updates: Dict[str, Dict[str, Any]],
-    config_path_override: str,
-) -> None:
-    """
-    Overwrite the existing config file with a new version that includes the desired `updates`.
-
-    :param updates: A nested `dict`, where top-level keys correspond to configuration sections and top-level values contain `dict`s whose key/value pairs correspond to the desired option/value updates.
-    :param config_path: The path to the current configuration file.
-    """
-
-    try:
-        config_path = config_path_override or load_file(
-            file_names=[DEFAULT_CONFIG_PATH]
-        )
-    except FileNotFoundError as error:
-        raise error
-
-    current_config = load(config_path)
-
-    for key, value in updates.items():
-        if key in current_config:
-            current_config[key].update(value)
-        else:
-            current_config.update({key: value})
-
-    with open(config_path, "w", encoding="utf-8") as config_file:
-        dump(current_config, config_file)
-
-    echo(f"Updated {config_path}:")
-
-    for key, value in updates.items():
-        for subkey, val in value.items():
-            echo(f"\tSet {key}.{subkey} = {val}")
