+++ tests/ops/fixtures/application_fixtures.py
@@ -55,7 +55,7 @@ from fides.api.ops.service.masking.strategy.masking_strategy_string_rewrite impo
 )
 from fides.api.ops.util.data_category import DataCategory
 from fides.ctl.core.config import get_config
-from fides.ctl.core.config.utils import load_file
+from fides.ctl.core.config.helpers import load_file
 from fides.lib.models.audit_log import AuditLog, AuditLogAction
 from fides.lib.models.client import ClientDetail
 from fides.lib.models.fides_user import FidesUser
