+++ tests/ops/service/connectors/test_connector_registry_service.py
@@ -12,7 +12,7 @@ from fides.api.ops.service.connectors.saas.connector_registry_service import (
     update_saas_configs,
 )
 from fides.api.ops.util.saas_util import load_config, load_dataset, load_yaml_as_string
-from fides.ctl.core.config.utils import load_file
+from fides.ctl.core.config.helpers import load_file
 
 NEW_CONFIG_DESCRIPTION = "new test config description"
 NEW_DATASET_DESCRIPTION = "new test dataset description"
